+++ dispatch/android/platform/adamma/c4dhi/claid_platform_impl/DeviceInfoGathererAndroid.java
@@ -20,164 +20,50 @@
  ***************************************************************************/
 package adamma.c4dhi.claid_platform_impl;
 
-import static android.app.AlarmManager.RTC_WAKEUP;
-import static android.app.PendingIntent.FLAG_MUTABLE;
-import static android.app.PendingIntent.FLAG_UPDATE_CURRENT;
-import static android.app.PendingIntent.getBroadcast;
-import static android.content.Context.ALARM_SERVICE;
-import static android.content.Context.POWER_SERVICE;
-
-import java.io.BufferedWriter;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.nio.Buffer;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.Calendar;
-
-import android.app.AlarmManager;
-import android.app.PendingIntent;
-import android.content.Context;
 import android.content.Intent;
+import android.content.IntentFilter;
+import android.os.BatteryManager;
 import android.os.PowerManager;
 
-import adamma.c4dhi.claid.Logger.Logger;
 import adamma.c4dhi.claid.RemoteFunction.RemoteFunctionRunnableHandler;
-import java.util.Date;
-import java.text.SimpleDateFormat;
-public class GlobalDeviceSchedulerAndroid
+
+public class DeviceInfoGathererAndroid
 {
-    private PowerManager.WakeLock wakeLock;
 
-    public GlobalDeviceSchedulerAndroid(RemoteFunctionRunnableHandler runnableHandler)
+    public DeviceInfoGathererAndroid(RemoteFunctionRunnableHandler runnableHandler)
     {
-        runnableHandler.registerRunnable(this, "android_acquire_wakelock", Void.class);
-        runnableHandler.registerRunnable(this, "android_release_wakelock", Void.class);
-        runnableHandler.registerRunnable(this, "android_schedule_device_wakeup", Void.class, Long.class);
+        runnableHandler.registerRunnable(this, "android_get_battery_level_percentage", Double.class);
+        runnableHandler.registerRunnable(this, "android_is_device_charging", Boolean.class);
     }
 
-    private boolean createWakeLocKIfNotExists()
+    public Double android_get_battery_level_percentage()
     {
-        Context context = CLAID.getContext();
-        if(context == null)
-        {
-            writeToLogFile("Cannot create WakeLock, context is null");
-            return false;
-        }
-        if(wakeLock != null)
-        {
-            return true;
-        }
-        PowerManager powerManager = (PowerManager) context.getSystemService(POWER_SERVICE);
-        wakeLock = powerManager.newWakeLock(PowerManager.PARTIAL_WAKE_LOCK,
-                "CLAID::GlobalDeviceSchedulerAndroid::PartialWakeLock");
-        writeToLogFile("Created WakeLock");
+        IntentFilter intentFilter = new IntentFilter(Intent.ACTION_BATTERY_CHANGED);
+        Intent batteryStatus = CLAID.getContext().registerReceiver(null, intentFilter);
 
-        return true;
+        if (batteryStatus != null) {
+            int level = batteryStatus.getIntExtra(BatteryManager.EXTRA_LEVEL, -1);
+            int scale = batteryStatus.getIntExtra(BatteryManager.EXTRA_SCALE, -1);
+
+            // Calculate battery percentage
+            double batteryPercentage = (level / (double) scale) * 100;
+            return batteryPercentage;
+        } else {
+            return -1.0;
+        }
     }
 
-    public void android_acquire_wakelock()
+    public Boolean android_is_device_charging()
     {
-        if(!createWakeLocKIfNotExists())
-        {
-            Logger.logError("Failed to acquire wake lock! Cannot create WakeLock object.");
-            return;
-        }
-        writeToLogFile("android_acquire_wakelock called");
+        IntentFilter intentFilter = new IntentFilter(Intent.ACTION_BATTERY_CHANGED);
+        Intent batteryStatus = CLAID.getContext().registerReceiver(null, intentFilter);
 
-        if(!wakeLock.isHeld())
-        {
-            writeToLogFile("android_acquire_wakelock acquired");
-            wakeLock.acquire();
-        }
-        else
-        {
-            writeToLogFile("android_acquire_wakelock do nothing because wake lock is already held");
+        if (batteryStatus != null) {
+            int status = batteryStatus.getIntExtra(BatteryManager.EXTRA_STATUS, -1);
+            return status == BatteryManager.BATTERY_STATUS_CHARGING;
+        } else {
+            return false; // Assume not charging if we can't retrieve status
         }
     }
 
-    public void android_release_wakelock()
-    {
-        if(!createWakeLocKIfNotExists())
-        {
-            Logger.logError("Failed to acquire wake lock! Cannot create WakeLock object.");
-            return;
-        }
-        writeToLogFile("android_release_wakelock called");
-        if(wakeLock.isHeld())
-        {
-            writeToLogFile("android_release_wakelock releasing wakelock");
-            wakeLock.release();
-        }
-    }
-
-    public void android_schedule_device_wakeup(Long scheduledTimeUnixTimestamp)
-    {
-        Logger.logInfo("android_schedule_device_wakeup_at called");
-        writeToLogFile("android_schedule_device_wakeup_at called");
-
-        Context context = CLAID.getContext();
-
-        if(context == null)
-        {
-            Logger.logError("Failed to schedule device wake up at time. CLAID context is null.");
-            return;
-        }
- 
-        Calendar calendar = Calendar.getInstance();
-
-        writeToLogFile("android_schedule_device_wakeup_at scheduling");
-
-        long milliseconds = scheduledTimeUnixTimestamp.longValue();
-        calendar.setTimeInMillis(milliseconds);
-
-        final AlarmManager am = (AlarmManager) context.getSystemService(ALARM_SERVICE);
-        final Intent intent = new Intent(context, adamma.c4dhi.claid_platform_impl.GlobalDeviceSchedulerWakeUpReceiver.class);
-        final PendingIntent pending = getBroadcast(context, 0, intent, FLAG_UPDATE_CURRENT|FLAG_MUTABLE);
-
-        Logger.logInfo("Scheduling device wakeup at: " + calendar.toString());
-
-        // Convert Calendar to Date
-        Date date = calendar.getTime();
-        // Define the desired date format
-        SimpleDateFormat dateFormat = new SimpleDateFormat("dd-MM-yyyy_HH:mm:ss");
-        // Format the date
-        String formattedDate = dateFormat.format(date);
-
-   
-        writeToLogFile("Scheduling device wakeup at: " + formattedDate);
-
-        am.setExactAndAllowWhileIdle(RTC_WAKEUP, calendar.getTimeInMillis(), pending);
-    }
-
-    private BufferedWriter logWriter = null;
-    private void writeToLogFile(String text) {
-        try {
-            // Get the current date and time
-            LocalDateTime now = LocalDateTime.now();
-            DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd-MM-yyyy_HH-mm-ss");
-            String formattedDateTime = now.format(formatter);
-            if(logWriter == null)
-            {
-                // Open the file in append mode
-                String filename = CLAID.getCommonDataPath() + "/scheduler_log-"  + formattedDateTime + ".txt";
-                Logger.logInfo("Opening scheduler log " + filename);
-                logWriter = new BufferedWriter(new FileWriter(filename, true));
-            }
-
-            // Get the Unix timestamp
-            long unixTimestamp = System.currentTimeMillis();
-
-            // Prepend the date, time, and Unix timestamp to the text
-            String logEntry = formattedDateTime + " [" + unixTimestamp + "] " + text;
-
-            // Write the text to the file
-            logWriter.write(logEntry);
-            // New line for better readability in the log file
-            logWriter.newLine();
-            logWriter.flush();
-        } catch (IOException e) {
-            e.printStackTrace();
-        } 
-    }
 }